chore: remove test web components
