+++ packages/fuel-indexer-tests/assets/bin/fuel-node.d
@@ -1,0 +1,1 @@
+/Users/rashad/development/repos/fuel-indexer/target/release/fuel-node: /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer/src/database.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer/src/executor.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer/src/ffi.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer/src/service.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-api-server/src/api.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-api-server/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-api-server/src/uses.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/database-types/src/directives.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/database-types/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/postgres/migrations/20220713212919_initial_migration.down.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/postgres/migrations/20220713212919_initial_migration.up.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/postgres/migrations/2022090908251_create_asset_registry.down.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/postgres/migrations/2022090908251_create_asset_registry.up.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/postgres/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/sqlite/migrations/20220713212919_initial_migration.down.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/sqlite/migrations/20220713212919_initial_migration.up.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/sqlite/migrations/2022090908251_create_asset_registry.down.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/sqlite/migrations/2022090908251_create_asset_registry.up.sql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/sqlite/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-database/src/queries.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/config/database.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/config/fuel_node.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/config/graphql.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/config/mod.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/defaults.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/manifest.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-lib/src/utils.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-metrics/src/core.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-metrics/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-metrics/src/queries/mod.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-metrics/src/queries/postgres.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-metrics/src/queries/sqlite.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/base.graphql /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/db/graphql.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/db/manager.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/db/tables.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/db.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-schema/src/utils.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/assets/bad_simple_wasm.wasm /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/assets/bad_simple_wasm.yaml /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/assets/fuel_indexer_test.yaml /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/assets/simple_wasm.wasm /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/assets/simple_wasm.yaml /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/components/web/src/bin/fuel-node.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/src/fixtures.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-tests/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-types/src/abi.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-types/src/ffi.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-types/src/graphql.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-types/src/lib.rs /Users/rashad/development/repos/fuel-indexer/packages/fuel-indexer-types/src/tx.rs
