+++ packages/fuel-indexer-tests/components/fuel-node/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "fuel-node"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+clap = { version = "3.1", features = ["cargo", "derive", "env"] }
+fuel-indexer-tests = { version = "0.0.0", path = "./../../../fuel-indexer-tests" }
+fuel-types = "0.5"
+fuels = { version = "0.32", features = ["fuel-core-lib"] }
+fuels-abigen-macro = "0.32"
+tokio = { version = "1.17", features = ["macros", "rt-multi-thread"] }
