+++ packages/fuel-indexer-tests/components/web-api/src/main.rs
@@ -1,0 +1,7 @@
+use fuel_indexer_tests::fixtures::test_web::server;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let _ = server().await;
+    Ok(())
+}
