--- packages/fuel-indexer-tests/components/web/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "fuel-indexer-test-web"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[[bin]]
-name = "web-api"
-path = "src/bin/web-api.rs"
-
-[[bin]]
-name = "fuel-node"
-path = "src/bin/fuel-node.rs"
-
-[dependencies]
-actix-web = { version = "4", default-features = false }
-anyhow = { version = "1.0", default-features = false }
-async-std = "1"
-clap = { version = "3.1", features = ["cargo", "derive", "env"] }
-fuel-indexer-lib = { version = "0.1", path = "../../../fuel-indexer-lib" }
-fuel-indexer-tests = { version = "0.0.0", path = "./../../../fuel-indexer-tests" }
-fuel-tx = "0.23"
-fuel-types = "0.5"
-fuels = { version = "0.32", features = ["fuel-core-lib"] }
-fuels-abigen-macro = "0.32"
-fuels-core = { version = "0.32", default-features = false }
-rand = "0.8"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0", features = ["raw_value"] }
-thiserror = "1.0"
-tokio = { version = "1.17", features = ["macros", "rt-multi-thread"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.2", features = ["env-filter"] }
