--- packages/fuel-indexer-tests/components/web/src/bin/web-api.rs
@@ -1,7 +1,0 @@
-use fuel_indexer_tests::fixtures::test_web::server;
-
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let _ = server().await;
-    Ok(())
-}
