--- packages/fuel-indexer-tests/components/web/wallet.json
@@ -1,1 +1,0 @@
-{"crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"0cab63dc2ad04e6d0f24e5827c53bc7f"},"ciphertext":"2cb32369b2b65061faaa4d8dc976c0c1afaaf565064a7eeb2acbc1d3a5868edc","kdf":"scrypt","kdfparams":{"dklen":32,"n":8192,"p":1,"r":8,"salt":"5eb6e501b73dc0ab32949fca00616393471bc18691e76ed156f97818421fd7d2"},"mac":"e7094f9cb2a8ae9774e4e7aa9608be1cda6df8cbe14e46f584e708c7a99d3b3e"},"id":"4a7075a7-a84b-4205-92d9-d0f0e509134b","version":3}
